adding basket

+++ client/src/contexts/authContext.jsx
@@ -1,6 +1,5 @@
 
 import * as autService from '../api/autService'
-import * as Path from '../paths/paths'
 import {  useNavigate } from 'react-router-dom'
 import { createContext } from "react";
 import usePersistedState from '../hooks/usePersistedState';
@@ -12,19 +11,28 @@ import usePersistedState from '../hooks/usePersistedState';
     const [auth, setAuth] = usePersistedState('auth',{});
   
     const loginSubmitHandler = async (values) => {
+      try{
         const result = await autService.login(values.email, values.password);
 
         setAuth(result);
         localStorage.setItem('accessToken',result.accessToken);
         navigate('/');
+      }catch(err){
+        alert(err.message);
+      }
     };
 
      const registerSubmitHandler = async(values) =>{
 
+       if(values.password === values.confirmPassword){
        const result = await autService.register(values.username,values.password,values.email,values.image,values.firstName,values.lastName,values.age);
        setAuth(result);
        localStorage.setItem('accessToken',result.accessToken);
         navigate('/');
+       }else{
+        alert("pass and repass do not match") 
+       }
+      
      }
      
    const logoutHandler = () =>{
